Move SettingsActivity and SettingsFragment to preferences.ui namespace
