+++ app/src/main/java/org/jellyfin/androidtv/preferences/ui/SettingsFragment.java
@@ -1,4 +1,4 @@
-package org.jellyfin.androidtv.settings;
+package org.jellyfin.androidtv.preferences.ui;
 
 import android.app.AlertDialog;
 import android.content.SharedPreferences;
